requirements.txt
certifi==2022.9.24
click==8.1.3
cloudinary==1.30.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
python-decouple==3.6
six==1.16.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
urllib3==1.26.13
Werkzeug==2.2.2
